test(vats): makeFakeStorageKit helper
