+++ packages/vats/jsconfig.json
@@ -20,5 +20,6 @@
     "scripts/**/*.js",
     "src/**/*.js",
     "test/**/*.js",
+    "tools/**/*.js",
   ],
 }
